do not fail on existing link

+++ Rakefile.rb
@@ -66,7 +66,7 @@ namespace "prep" do
 
   static_graphs.each do |graph|
     file graph + "lower_bound" => graph do
-      sh "ln -s #{graph}/travel_time #{graph}/lower_bound"
+      sh "ln -s #{graph}/travel_time #{graph}/lower_bound" unless File.exist? "#{graph}/travel_time"
     end
   end
